Playing around with search result analysis
Trying to build a ML model that searches a phrase (specifically 'cannabis norge') and determines if site is positvly, negativly or neutrally skewed toward cannabis. So far, I gather the search results, and can analyze top 50 words of each. Next step is to store a list of top 50 for n websites, and then manually determine the polarity of each site. This will then be splt into a training and testing set.

This is a huge project.